# docs/docs_swagger/swagger.yaml
basePath: /api
...
host: localhost:10031
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: 智能门禁管理系统API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
...
  /admin/advertisement-building/advertisements:
    get:
      consumes:
      - application/json
      description: 根据建筑ID获取所有与之绑定的广告信息
      parameters:
      - description: 建筑ID
        in: query
        name: buildingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 广告列表信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的建筑ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 建筑不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取建筑绑定的广告
      tags:
      - Advertisement-Building
  /admin/advertisement-building/bind:
    post:
      consumes:
      - application/json
      description: 将一个或多个广告绑定到一个或多个建筑物
      parameters:
      - description: 绑定信息
        in: body
        name: bindingInfo
        required: true
        schema:
          type: object
      - description: 广告ID列表
        in: body
        name: advertisementIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: 建筑ID列表
        in: body
        name: buildingIds
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 绑定结果信息，包含成功绑定的关系和已存在的绑定
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 输入参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 广告或建筑不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 绑定广告到建筑
      tags:
      - Advertisement-Building
  /admin/advertisement-building/buildings:
    get:
      consumes:
      - application/json
      description: 根据广告ID获取所有与之绑定的建筑物信息
      parameters:
      - description: 广告ID
        in: query
        name: advertisementId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 建筑列表信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的广告ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 广告不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取广告绑定的建筑
      tags:
      - Advertisement-Building
  /admin/advertisement-building/unbind:
    post:
      consumes:
      - application/json
      description: 解除一个广告与多个建筑物的绑定关系
      parameters:
      - description: 解绑信息
        in: body
        name: unbindInfo
        required: true
        schema:
          type: object
      - description: 广告ID
        in: body
        name: advertisementId
        required: true
        schema:
          type: integer
      - description: 建筑ID列表
        in: body
        name: buildingIds
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 解绑成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 输入参数错误或建筑未绑定
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 广告或建筑不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 解绑广告与建筑
      tags:
      - Advertisement-Building
...
  /building_admin/login:
    post:
      consumes:
      - application/json
      description: 建筑管理员通过邮箱和密码登录系统
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          type: object
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 包含token和登录成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 登录失败信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 认证失败
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - None: []
      summary: 建筑管理员登录
      tags:
      - BuildingAdmin
...
securityDefinitions:
  BearerAuth:
    description: 输入带有Bearer前缀的用户认证token
    in: header
    name: Authorization
    type: apiKey
  DeviceAuth:
    description: 输入带有Bearer前缀的设备认证token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"